+++ src/main/java/wooteco/subway/service/member/MemberService.java
@@ -1,20 +1,10 @@
 package wooteco.subway.service.member;
 
-import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.stream.Collectors;
-
 import org.springframework.stereotype.Service;
 
-import wooteco.subway.domain.member.Favorite;
-import wooteco.subway.domain.member.Favorites;
 import wooteco.subway.domain.member.Member;
 import wooteco.subway.domain.member.MemberRepository;
-import wooteco.subway.domain.station.Station;
-import wooteco.subway.domain.station.StationRepository;
 import wooteco.subway.infra.JwtTokenProvider;
-import wooteco.subway.service.member.dto.FavoriteRequest;
-import wooteco.subway.service.member.dto.FavoriteResponse;
 import wooteco.subway.service.member.dto.LoginRequest;
 import wooteco.subway.service.member.dto.MemberResponse;
 import wooteco.subway.service.member.dto.UpdateMemberRequest;
@@ -23,13 +13,10 @@ import wooteco.subway.web.exceptions.MemberNotFoundException;
 
 @Service
 public class MemberService {
-    private final StationRepository stationRepository;
     private final MemberRepository memberRepository;
     private final JwtTokenProvider jwtTokenProvider;
 
-    public MemberService(final StationRepository stationRepository,
-        final MemberRepository memberRepository, final JwtTokenProvider jwtTokenProvider) {
-        this.stationRepository = stationRepository;
+    public MemberService(MemberRepository memberRepository, JwtTokenProvider jwtTokenProvider) {
         this.memberRepository = memberRepository;
         this.jwtTokenProvider = jwtTokenProvider;
     }
@@ -58,30 +45,6 @@ public class MemberService {
 
     public Member findMemberByEmail(String email) {
         return memberRepository.findByEmail(email)
-            .orElseThrow(MemberNotFoundException::new);
-    }
-
-    public void addFavorite(final Member member, final FavoriteRequest request) {
-        Station source = stationRepository.findById(request.getSource())
-            .orElseThrow(NoSuchElementException::new);
-        Station target = stationRepository.findById(request.getTarget())
-            .orElseThrow(NoSuchElementException::new);
-        member.addFavorite(new Favorite(source.getId(), target.getId()));
-        memberRepository.save(member);
-    }
-
-    public List<FavoriteResponse> getAllFavorites(final Member member) {
-        Favorites favorites = member.getFavorites();
-        List<Station> stations = stationRepository.findAllById(favorites.getAllStations());
-
-        return favorites.getFavorites()
-            .stream()
-            .map(favorite -> FavoriteResponse.of(favorite, stations))
-            .collect(Collectors.toList());
-    }
-
-    public void removeFavoriteById(final Member member, final Long id) {
-        member.removeFavorite(id);
-        memberRepository.save(member);
+                .orElseThrow(MemberNotFoundException::new);
     }
 }
