+++ src/main/java/wooteco/subway/web/member/FavoriteController.java
@@ -12,34 +12,34 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import wooteco.subway.domain.member.Member;
-import wooteco.subway.service.member.MemberService;
+import wooteco.subway.service.member.FavoriteService;
 import wooteco.subway.service.member.dto.FavoriteRequest;
 import wooteco.subway.service.member.dto.FavoriteResponse;
 
 @RequestMapping("/favorites")
 @RestController
 public class FavoriteController {
-    private final MemberService memberService;
+    private final FavoriteService favoriteService;
 
-    public FavoriteController(final MemberService memberService) {
-        this.memberService = memberService;
+    public FavoriteController(FavoriteService favoriteService) {
+        this.favoriteService = favoriteService;
     }
 
     @PostMapping
     public ResponseEntity<Void> create(@LoginMember Member member, @RequestBody FavoriteRequest request) {
-        memberService.addFavorite(member, request);
+        favoriteService.addFavorite(member, request);
         return ResponseEntity.ok().build();
     }
 
     @GetMapping
     public ResponseEntity<List<FavoriteResponse>> getAll(@LoginMember Member member) {
-        List<FavoriteResponse> response = memberService.getAllFavorites(member);
+        List<FavoriteResponse> response = favoriteService.getAllFavorites(member);
         return ResponseEntity.ok(response);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> remove(@LoginMember Member member, @PathVariable Long id) {
-        memberService.removeFavoriteById(member, id);
+        favoriteService.removeFavoriteById(member, id);
         return ResponseEntity.noContent().build();
     }
 }
