+++ src/test/java/wooteco/subway/service/member/MemberServiceTest.java
@@ -39,9 +39,6 @@ public class MemberServiceTest {
     private MemberService memberService;
 
     @Mock
-    private StationRepository stationRepository;
-
-    @Mock
     private MemberRepository memberRepository;
 
     @Mock
@@ -49,7 +46,7 @@ public class MemberServiceTest {
 
     @BeforeEach
     void setUp() {
-        this.memberService = new MemberService(stationRepository, memberRepository, jwtTokenProvider);
+        this.memberService = new MemberService(memberRepository, jwtTokenProvider);
     }
 
     @DisplayName("회원 생성 테스트")
@@ -134,64 +131,5 @@ public class MemberServiceTest {
 
         // then
         verify(memberRepository).deleteById(member.getId());
-    }
-
-    @DisplayName("즐겨찾기 생성 테스트")
-    @Test
-    void addFavorite() {
-        // given
-        Member member = new Member(TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
-        Station source = new Station(1L, "잠실역");
-        Station target = new Station(3L, "잠실역");
-        given(stationRepository.findById(source.getId())).willReturn(Optional.of(source));
-        given(stationRepository.findById(target.getId())).willReturn(Optional.of(target));
-        given(memberRepository.save(any())).willReturn(member);
-
-        // when
-        memberService.addFavorite(member, new FavoriteRequest(source.getId(), target.getId()));
-
-        // then
-        verify(memberRepository).save(any());
-    }
-
-    @DisplayName("즐겨찾기 목록 조회 테스트")
-    @Test
-    void getAllFavorites() {
-        // given
-        Member member = new Member(TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
-        Station source = new Station(1L, "잠실역");
-        Station source2 = new Station(2L, "삼성역");
-        Station target = new Station(3L, "석촌역");
-        Station target2 = new Station(4L, "선릉역");
-
-        given(stationRepository.findById(source.getId())).willReturn(Optional.of(source));
-        given(stationRepository.findById(source2.getId())).willReturn(Optional.of(source2));
-        given(stationRepository.findById(target.getId())).willReturn(Optional.of(target));
-        given(stationRepository.findById(target2.getId())).willReturn(Optional.of(target2));
-        given(memberRepository.save(any())).willReturn(member);
-        given(stationRepository.findAllById(anyList())).willReturn(Arrays.asList(source, target, source2, target2));
-
-        memberService.addFavorite(member, new FavoriteRequest(source.getId(), target.getId()));
-        memberService.addFavorite(member, new FavoriteRequest(source2.getId(), target2.getId()));
-
-        // when
-        List<FavoriteResponse> favorites = memberService.getAllFavorites(member);
-
-        // then
-        assertThat(favorites).hasSize(2);
-    }
-
-    @DisplayName("즐겨찾기 삭제 테스트")
-    @Test
-    void deleteFavorite() {
-        // given
-        Member member = new Member(TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
-        member.addFavorite(new Favorite(1L, 1L, 2L));
-
-        // when
-        memberService.removeFavoriteById(member, 1L);
-
-        // then
-        verify(memberRepository).save(any());
     }
 }
